Dependency Injection using @Autowired annotation
Here, the required dependency is the FortuneService and its interface is implemented using HappyFortuneService class.

A constructor is created in TennisCoach class for the required dependency to be injected and it is marked with @Autowire annotation so that Spring can scan for a component that implements FortuneService interface which is the HappyFortuneService class (it gets scanned since it is marked with @Component annotation). Then, the required dependency gets injected.