@Bean annotation
The @Bean annotation is used to directly create a bean in the config file where the method is the bean_id.

Here, FortuneService interface is implemented by SadFortuneService class and the @Component is not marked in the class. So, even the @ComponentScan annotation is also not needed to scan for the components as the beans are directly created in the config file using @Bean annotation.